backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
anthropic
yt-dlp
google-api-python-client
apscheduler
pydantic
beautifulsoup4
pymupdf
